Extract parseJson into MerlinParsers
